pyproject.toml
[project]
name = "civilwar-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

